♻️(project-domain) move ddd object in a separate lib in monorepo
