+++ libs/ddd/aggregate_root.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass, field, replace
 from functools import wraps
 from typing import Callable
 
-from domain.ddd.domain_event import DomainEvent, DomainEventMetadata
-from domain.ddd.entity import Entity
+from ddd.domain_event import DomainEvent, DomainEventMetadata
+from ddd.entity import Entity
 
 
 @dataclass(kw_only=True)
