+++ libs/ddd/entity.py
