+++ libs/ddd/entity_interface.py
