+++ libs/ddd/tests/example_aggregate_factory.py
@@ -1,0 +1,7 @@
+from tests.example_aggregate import ExampleAggregate
+
+
+class ExampleAggregateFactory:
+    @staticmethod
+    def build(name: str = "Example") -> ExampleAggregate:
+        return ExampleAggregate.build(name=name)
