+++ src/web/application/ingestion/usecases/archive_offers.py
@@ -1,8 +1,8 @@
 from datetime import datetime
 
 from asgiref.sync import async_to_sync
+from ddd.usecase_interface import IUseCase
 
-from domain.ddd.usecase_interface import IUseCase
 from domain.gateways.document_gateway_interface import IDocumentGateway
 from domain.repositories.offers_repository_interface import (
     IArchiveResult,
