+++ src/web/application/ingestion/usecases/load_offers.py
@@ -1,9 +1,9 @@
 from typing import List, cast
 
 from asgiref.sync import sync_to_async
+from ddd.async_usecase_interface import IAsyncUseCase
 
 from application.ingestion.interfaces.load_documents_input import LoadDocumentsInput
-from domain.ddd.async_usecase_interface import IAsyncUseCase
 from domain.entities.document import DocumentType
 from domain.exceptions.document_error import InvalidDocumentTypeError
 from domain.gateways.document_gateway_interface import IDocumentGateway
