+++ src/web/application/ingestion/usecases/upsert_offers.py
@@ -1,7 +1,8 @@
+from ddd.usecase_interface import IUseCase
+
 from application.ingestion.interfaces.upsert_offers_input import (
     UpsertOffersInput,
 )
-from domain.ddd.usecase_interface import IUseCase
 from domain.repositories.document_repository_interface import IUpsertResult
 from domain.repositories.offers_repository_interface import IOffersRepository
 from domain.services.logger_interface import ILogger
