+++ src/web/domain/candidature/events/candidature_events.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
 from uuid import UUID
 
-from domain.ddd.domain_event import DomainEvent
+from ddd.domain_event import DomainEvent
+
 from domain.shared.value_objects.etapes_recrutement import EtapeRecrutement
