+++ src/web/domain/repositories/offers_repository_interface.py
@@ -1,7 +1,8 @@
 from typing import List, Protocol, TypedDict
 from uuid import UUID
 
-from domain.ddd.page_interface import IPage
+from ddd.page_interface import IPage
+
 from domain.entities.offer import Offer
 from domain.repositories.document_repository_interface import IUpsertResult
