+++ src/web/domain/services/text_extractor_interface.py
@@ -2,7 +2,8 @@
 
 from typing import Any, Dict, Protocol, Union
 
-from domain.ddd.entity_interface import IEntity
+from ddd.entity_interface import IEntity
+
 from domain.entities.document import Document
