+++ src/web/infrastructure/di/ingestion/ingestion_container.py
@@ -1,3 +1,5 @@
+from ddd.async_usecase_interface import IAsyncUseCase
+from ddd.entity_interface import IEntity
 from dependency_injector import containers, providers
 
 from application.ingestion.interfaces.load_documents_input import LoadDocumentsInput
@@ -13,8 +15,6 @@ from application.ingestion.usecases.load_documents import LoadDocumentsUsecase
 from application.ingestion.usecases.load_offers import LoadOffersUsecase
 from application.ingestion.usecases.upsert_offers import UpsertOffersUseCase
 from application.ingestion.usecases.vectorize_documents import VectorizeDocumentsUsecase
-from domain.ddd.async_usecase_interface import IAsyncUseCase
-from domain.ddd.entity_interface import IEntity
 from domain.repositories.document_repository_interface import (
     IUpsertResult,
 )
