+++ src/web/infrastructure/repositories/shared/postgres_metier_repository.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 from typing import Dict, List
 
+from ddd.page_interface import IPage
 from django.db import DatabaseError, transaction
 from django.db.models import F, Q
 from django.utils import timezone
 
-from domain.ddd.page_interface import IPage
 from domain.entities.metier import Metier
 from domain.entities.offer import Offer
 from domain.exceptions.metiers_error import MetierDoesNotExist
