+++ src/web/presentation/ingestion/mappers.py
@@ -1,6 +1,7 @@
 from typing import Optional
 
-from domain.ddd.mapper_interface import IToDomainMapper
+from ddd.mapper_interface import IToDomainMapper
+
 from domain.entities.offer import Offer
 from domain.value_objects.area import GeographicalArea
 from domain.value_objects.category import Category
