+++ src/web/pyproject.toml
@@ -5,6 +5,7 @@ description = "The CSPLab Monolith"
 readme = "README.md"
 requires-python = "~=3.12.0"
 dependencies = [
+    "ddd",
     "dependency-injector>=4.48.2,<4.49.0",
     "django>=6.0.1,<7.0.0",
     "django-dsfr>=3.3.0,<3.4.0",
@@ -67,6 +68,7 @@ url = "https://pypi.org/simple"
 
 [tool.uv.sources]
 web = { path = ".", editable = true }
+ddd = { path = "../../libs/ddd", editable = true }
 
 [tool.hatch.build.targets.wheel]
 packages = [
